chore: move method handlers to method channel class
